requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
redis==5.0.1
pinecone-client==3.0.2
numpy==1.24.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
pydantic==2.5.3
langchain==0.1.0
langchain-pinecone==0.0.1 